Framework: attempt to read the input files directly from disk
Rather than from the resource, which could be in a zipped .jar file.

This brings the Day 1 parsing: 0.27 -> 0.21 ms, but this still seems
very high for a ~1kB file.

After further investigation, it seems it is the call to
'new FileInputStream' itself that takes significant time (~60-100 us)
with quite a lot of variance. I think it would be good to isolate
this out from other measurements.