Framework: preload files/resources into a byte[]
This moves the variable loading time out of the measurements, while
keeping the 'parse' logic of the puzzle measured.

This will make for a fairer comparison of the 'hot' Java code with
other language implementations (which may also not include this),
and eliminate effects of IO latency and caching etc.