Day 4 - change to byte biting
Again the key to high speed appears to be less about dynamic memory
(not close to the top contributor yet) but instead about reducing
String construction. The input is processed as bytes (to avoid
char-encoding) and mostly parsed in one phase. (Placing in a List
is not significant either, yet.)

We avoid creating, and switching on, String by reinterpreting the
4 character key (including the ':') as an int from the binary stream.

Perf: 505 -> 242 us (laptop)