2023 Day 12 - ThreadLocal, custom-size caches (8 -> 6 ms)
Avoid re-allocating cache arrays by making them thread-local and
just reset them in-place.

Using a smaller cache for part 1 makes that significantly faster
(1300 > 300 us).

The next biggest gains would be in the parsing code, but that's less
interesting - I optimised that sort of thing a few times already
back in 2020.